requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.13.4
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
networkx==3.1
numpy==1.24.4
packaging==24.0
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyobjc-core==10.2
pyobjc-framework-ApplicationServices==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-CoreText==10.2
pyobjc-framework-Quartz==10.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rubicon-objc==0.4.8
six==1.16.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.2
transformers==4.28.1
typing_extensions==4.11.0
urllib3==2.2.1
